Build Mode to Blueprint Mode bug fix
When the player switches from Build Mode to Blueprint Mode at the start of a game, the Build Tool wouldn't unequip and throws an error.

Note: Made changes to the Inventory/Quests/Builder UI system so they are always active but hidden via CanvasGroups to keep their scripts running. This fixes the error mentioned above.