Reverted Script Execution Order
In the previous commit, I made changes to the scripts execution order by moving all scripts besides the SaveLoadController before the Default Time in an attempt to make the Chest load in all the saved Items. It did not work and prevented the player's inventory Items to load in (UI). This commit brings the project back to working order, besides the issue where the chest Items won't save or load.